LDA
This file contains the lda technique applied on the BCI dataset of 31784 rows and 11 columns, regression scores have been compared before and after of applying LDA , Also we have found the minimum number of components required for having 95% variance in the dataset. Regression score have been predicted in this too.